requirements.txt
#requirements.txt
flask
flask-sqlalchemy
psycopg2-binary  # For PostgreSQL
flask-login
Flask-Migrate
passlib
python-dotenv
PyJWT
Flask-Mail
itsdangerous
